backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic-settings==2.7.1
openai-whisper==20240930
ffmpeg-python==0.2.0
sse-starlette==2.2.1
python-dotenv==1.0.1
pytest==8.3.4
